requirements.txt
Appium-Python-Client
pytest
pytest-xdist
allure-pytest